requirements.txt
valohai-utils
tensorflow-datasets